Make 'terraform state pull > foo.json' work better
Before, info and output messages would be printed to STDOUT, and running
'terraform init' before 'terraform state pull' would also go to STDOUT.

This change forces all the non-error "echo"s to go to STDERR,
and it changes the 'terraform init' run before 'terraform state' to
output to STDERR.

This may cause an inconsistency bug when handling other commands
(the 'terraform init' output will go to a different FD for every other command)
so this may need to change, or other functionality may need to change.

The first thought I have is that we can do something complicated, like
redirect the output of the dependency-running commands to a temporary
file and 'tail -f' the file to STDERR, until an error actually happens,
and then re-output it to STDOUT... which would allow preserving the
output of all commands to STDOUT. But that would then present the problem
that we'd have to do that for both STDERR and STDOUT for the dependent
command, but we wouldn't be able to interleave them back together
correctly on output. So we could just collapse the STDERR and STDOUT
of dependent commands together so re-outputting is interlaved correctly.
But then when an error does happen, the output of the command is going
to show up twice: once while we were 'tail -f'ing the command to STDERR,
and once after the error when we re-output everything to STDOUT.
So maybe the best solution is just to hope that the user captures STDERR
and just 'tail -f' to STDERR and the user has to just look through STDERR
for the output from the failed command.