Fix location of default plan file
I'm not sure when this happened, but the location of the plan file changed from
the current directory you run terraformsh from, to the root module directory.

I chose to put it in the current directory so that it's both easier to identify
plan files when they're created, and to keep the module directory free of
extraneous/temporary files.

This will be a breaking change for those who expect the previous behavior
(plan files going into the module directory)