Fix passing arbitrary options to commands
Before, terraformsh was supposed to pass arbitrary command-line options to the
terraform command if they didn't match a known terraform command name. However,
the functions were not actually using the "$@" being passed to them with the
options. This fixes the commands to use "$@".

For example, now it should correctly use the following:

  terraformsh plan -compact-warnings