Hack to fix apply/destroy
Agh, this needs to be refactored !

The 'apply' and 'destory' commands are being passed a plan file path
in order to apply/destroy the plan. But this conflicts with passing
the '$@' (which comes from ${OPTS[*]}) at the very end of the command,
because you get 'terraform apply planfile -some-options' which Terraform
breaks on.

So this fix changes around the order of arguments *only* for apply
and destroy, so they get '$@' ($OPTS) *before* the $APPLY_ARGS
or $DESTROY_ARGS arrays.

This introduces (yet another) inconsistency which needs to be solved!