Refactor argument parsing, add 'workspace' wrapper
This code changes how Terraform commands and options are parsed on the command-line.
Basically, now we use arrays internally to separate each command and its
options.

Before, you could only do this:

  terraformsh init plan apply -compact-warnings

And terraformsh would either pass those options to every command, or
pass them only to the very last command.

But you can now do this:

  terradformsh init -compact-warnings plan apply

And the '-compact-warnings' option will only go to 'init' command.

---

In addition, i've added a wrapper for the 'workspace' command.
I suspect there's many more commands that will need wrappers, and
I probably need to go through and put more logic in to differentiate
from a root-level command, and sub-commands that are passed as options
to root-level commands.