Change behavior of cleanup of temp dir
Before, if an error occurred, removal of the temporary directory used by
TF_DATA_DIR would not happen (so you could inspect the failure).

This leads to the unfortunate problem that old state will get re-used by
a later run of Terraform, which you probably don't want in most cases.

Therefore, here we're only avoiding cleanup on error if the
NO_CLEANUP_TMP_ON_ERROR environment variable is set to '1'.