Bug fix: detect apply plan argument (#35)
If 'terraformsh apply' is run, by default we pass variable files, since apply
might want inputs.

But if the user passes 'terraformsh apply PLANFILE', then Terraform cannot accept
any variable file inputs (because the plan is already set).

This code will detect if the user passes any arguments after 'terraformsh apply'
that do not begin with '-' (as in -some-option). If any such argument exists and
is a file, this is assumed to be a plan file, and so the variable files will not
be passed.

Co-authored-by: Peter Willis <pwillis@outsideinc.com>